rename newqueryviewer
