+++ packages/builder/src/pages/builder/app/[application]/data/query/[queryId]/index.svelte
@@ -1,6 +1,6 @@
 <script>
   import { database, queries, datasources } from "stores/backend"
-  import QueryViewer from "components/integration/NewQueryViewer.svelte"
+  import QueryViewer from "components/integration/QueryViewer.svelte"
   import RestQueryViewer from "components/integration/RestQueryViewer.svelte"
   import { IntegrationTypes } from "constants/backend"
   import { cloneDeep } from "lodash/fp"
